Moving shared pieces to core
