+++ x-pack/plugin/core/src/main/java/org/elasticsearch/xpack/core/ml/utils/InferenceProcessorInfoExtractor.java
@@ -5,8 +5,9 @@
  * 2.0.
  */
 
-package org.elasticsearch.xpack.inference.utils;
+package org.elasticsearch.xpack.core.ml.utils;
 
+import org.apache.lucene.util.Counter;
 import org.elasticsearch.cluster.ClusterState;
 import org.elasticsearch.cluster.metadata.Metadata;
 import org.elasticsearch.ingest.ConfigurationUtils;
@@ -14,6 +15,7 @@ import org.elasticsearch.ingest.IngestMetadata;
 import org.elasticsearch.ingest.Pipeline;
 import org.elasticsearch.transport.Transports;
 
+import java.util.HashMap;
 import java.util.HashSet;
 import java.util.LinkedHashSet;
 import java.util.List;
@@ -26,13 +28,8 @@ import static org.elasticsearch.ingest.Pipeline.PROCESSORS_KEY;
 
 /**
  * Utilities for extracting information around inference processors from IngestMetadata
- * <p>
- * this class was duplicated from org.elasticsearch.xpack.ml.utils.InferenceProcessorInfoExtractor
  */
-
 public final class InferenceProcessorInfoExtractor {
-
-    public static final String TYPE = "inference";
     private static final String FOREACH_PROCESSOR_NAME = "foreach";
     // Any more than 10 nestings of processors, we stop searching for inference processor definitions
     private static final int MAX_INFERENCE_PROCESSOR_SEARCH_RECURSIONS = 10;
@@ -40,7 +37,40 @@ public final class InferenceProcessorInfoExtractor {
     private InferenceProcessorInfoExtractor() {}
 
     /**
-     * @param ingestMetadata The ingestMetadata of current {@link ClusterState}
+     * @param state The current cluster state
+     * @return The current count of inference processors
+     */
+    @SuppressWarnings("unchecked")
+    public static int countInferenceProcessors(ClusterState state) {
+        Metadata metadata = state.getMetadata();
+        if (metadata == null) {
+            return 0;
+        }
+        IngestMetadata ingestMetadata = metadata.custom(IngestMetadata.TYPE);
+        if (ingestMetadata == null) {
+            return 0;
+        }
+        Counter counter = Counter.newCounter();
+        ingestMetadata.getPipelines().forEach((pipelineId, configuration) -> {
+            Map<String, Object> configMap = configuration.getConfigAsMap();
+            List<Map<String, Object>> processorConfigs = ConfigurationUtils.readList(null, null, configMap, PROCESSORS_KEY);
+            for (Map<String, Object> processorConfigWithKey : processorConfigs) {
+                for (Map.Entry<String, Object> entry : processorConfigWithKey.entrySet()) {
+                    addModelsAndPipelines(
+                        entry.getKey(),
+                        pipelineId,
+                        (Map<String, Object>) entry.getValue(),
+                        pam -> counter.addAndGet(1),
+                        0
+                    );
+                }
+            }
+        });
+        return (int) counter.get();
+    }
+
+    /**
+     * @param ingestMetadata The ingestMetadata of current ClusterState
      * @return The set of model IDs referenced by inference processors
      */
     @SuppressWarnings("unchecked")
@@ -66,6 +96,38 @@ public final class InferenceProcessorInfoExtractor {
             }
         });
         return modelIds;
+    }
+
+    /**
+     * @param state Current cluster state
+     * @return a map from Model or Deployment IDs or Aliases to each pipeline referencing them.
+     */
+    @SuppressWarnings("unchecked")
+    public static Map<String, Set<String>> pipelineIdsByResource(ClusterState state, Set<String> ids) {
+        assert Transports.assertNotTransportThread("non-trivial nested loops over cluster state structures");
+        Map<String, Set<String>> pipelineIdsByModelIds = new HashMap<>();
+        Metadata metadata = state.metadata();
+        if (metadata == null) {
+            return pipelineIdsByModelIds;
+        }
+        IngestMetadata ingestMetadata = metadata.custom(IngestMetadata.TYPE);
+        if (ingestMetadata == null) {
+            return pipelineIdsByModelIds;
+        }
+        ingestMetadata.getPipelines().forEach((pipelineId, configuration) -> {
+            Map<String, Object> configMap = configuration.getConfigAsMap();
+            List<Map<String, Object>> processorConfigs = ConfigurationUtils.readList(null, null, configMap, PROCESSORS_KEY);
+            for (Map<String, Object> processorConfigWithKey : processorConfigs) {
+                for (Map.Entry<String, Object> entry : processorConfigWithKey.entrySet()) {
+                    addModelsAndPipelines(entry.getKey(), pipelineId, (Map<String, Object>) entry.getValue(), pam -> {
+                        if (ids.contains(pam.modelIdOrAlias)) {
+                            pipelineIdsByModelIds.computeIfAbsent(pam.modelIdOrAlias, m -> new LinkedHashSet<>()).add(pipelineId);
+                        }
+                    }, 0);
+                }
+            }
+        });
+        return pipelineIdsByModelIds;
     }
 
     /**
@@ -115,7 +177,7 @@ public final class InferenceProcessorInfoExtractor {
         if (processorType == null || processorDefinition == null) {
             return;
         }
-        if (TYPE.equals(processorType)) {
+        if (InferenceProcessorConstants.TYPE.equals(processorType)) {
             String modelId = (String) processorDefinition.get(MODEL_ID_RESULTS_FIELD);
             if (modelId != null) {
                 handler.accept(new PipelineAndModel(pipelineId, modelId));
